making InputStreamWrapper inner class for optimization
